+++ docs/docs/sandpack/files/table/code-tableValue.tsx
@@ -8,26 +8,26 @@ export const createTable = (): any => (
   <fragment>
     <htable colSizes={[150, 150, 150, 150]}>
       <htr>
-        <htd>
+        <hth>
           <hp>
             <htext />
           </hp>
-        </htd>
-        <htd>
+        </hth>
+        <hth>
           <hp>
             <htext bold>Human</htext>
           </hp>
-        </htd>
-        <htd>
+        </hth>
+        <hth>
           <hp>
             <htext bold>Dog</htext>
           </hp>
-        </htd>
-        <htd>
+        </hth>
+        <hth>
           <hp>
             <htext bold>Cat</htext>
           </hp>
-        </htd>
+        </hth>
       </htr>
       <htr>
         <htd>
@@ -64,12 +64,12 @@ export const createTable = (): any => (
     </htable>
   </fragment>
 );
-const attributes = { colspan: '2' };
+
 export const createSpanningTable = (): any => (
   <fragment>
     <htable colSizes={[300, 300]}>
       <htr>
-        <hth attributes={attributes}>
+        <hth colSpan={2}>
           <hp>
             <htext bold>Heading</htext>
           </hp>