+++ packages/nodes/indent/src/onKeyDownIndent.ts
@@ -12,6 +12,8 @@ export const onKeyDownIndent = <
 >(
   editor: E
 ): KeyboardHandlerReturnType => (e) => {
+  if (e.defaultPrevented) return;
+
   if (Hotkeys.isTab(editor, e)) {
     e.preventDefault();
     indent(editor);