+++ packages/nodes/table/src/components/TableCellElement/useTableCellElementState.ts
@@ -1,18 +1,27 @@
-import { TElement, useElement, usePlateEditorRef } from '@udecode/plate-common';
+import { useElement, usePlateEditorRef } from '@udecode/plate-common';
 import { useReadOnly } from 'slate-react';
-import { ELEMENT_TR } from '../../createTablePlugin';
-import { getTableColumnIndex } from '../../queries';
-import { useTableRowStore } from '../../stores/tableRowStore';
+import { ELEMENT_TABLE, ELEMENT_TR } from '../../createTablePlugin';
+import { getTableColumnIndex, getTableRowIndex } from '../../queries';
 import { useTableStore } from '../../stores/tableStore';
-import { TTableRowElement } from '../../types';
+import {
+  TTableCellElement,
+  TTableElement,
+  TTableRowElement,
+} from '../../types';
+import {
+  BorderStylesDefault,
+  getTableCellBorders,
+} from './getTableCellBorders';
 import { useIsCellSelected } from './useIsCellSelected';
 
 export type TableCellElementState = {
   colIndex: number;
+  rowIndex: number;
   readOnly: boolean;
   hovered: boolean;
   selected: boolean;
   rowSize: number | undefined;
+  borders: BorderStylesDefault;
 };
 
 export const useTableCellElementState = ({
@@ -24,23 +33,37 @@ export const useTableCellElementState = ({
   ignoreReadOnly?: boolean;
 } = {}): TableCellElementState => {
   const editor = usePlateEditorRef();
-  const cellElement = useElement<TElement>();
-  const hoveredColIndex = useTableStore().get.hoveredColIndex();
-  const isCellSelected = useIsCellSelected(cellElement);
+  const cellElement = useElement<TTableCellElement>();
 
-  const rowElement = useElement<TTableRowElement>(ELEMENT_TR);
-  const rowSizeOverride = useTableRowStore().get.overrideSize();
-  const rowSize = rowSizeOverride ?? rowElement?.size ?? undefined;
+  const colIndex = getTableColumnIndex(editor, cellElement);
+  const rowIndex = getTableRowIndex(editor, cellElement);
 
   const readOnly = useReadOnly();
 
-  const colIndex = getTableColumnIndex(editor, cellElement);
+  const isCellSelected = useIsCellSelected(cellElement);
+  const hoveredColIndex = useTableStore().get.hoveredColIndex();
+
+  const tableElement = useElement<TTableElement>(ELEMENT_TABLE);
+  const rowElement = useElement<TTableRowElement>(ELEMENT_TR);
+  const rowSizeOverrides = useTableStore().get.rowSizeOverrides();
+  const rowSize =
+    rowSizeOverrides.get(rowIndex) ?? rowElement?.size ?? undefined;
+
+  const isFirstCell = colIndex === 0;
+  const isFirstRow = tableElement.children[0] === rowElement;
+
+  const borders = getTableCellBorders(cellElement, {
+    isFirstCell,
+    isFirstRow,
+  });
 
   return {
     colIndex,
+    rowIndex,
     readOnly: !ignoreReadOnly && readOnly,
     selected: isCellSelected,
     hovered: hoveredColIndex === colIndex,
     rowSize,
+    borders,
   };
 };