+++ packages/nodes/table/src/queries/index.ts
@@ -4,6 +4,7 @@
 
 export * from './getCellInNextTableRow';
 export * from './getCellInPreviousTableRow';
+export * from './getLeftTableCell';
 export * from './getNextTableCell';
 export * from './getPreviousTableCell';
 export * from './getTableAbove';
@@ -12,3 +13,7 @@ export * from './getTableColumnIndex';
 export * from './getTableEntries';
 export * from './getTableGridAbove';
 export * from './getTableGridByRange';
+export * from './getTableOverriddenColSizes';
+export * from './getTableRowIndex';
+export * from './getTopTableCell';
+export * from './isTableBorderHidden';