+++ packages/ui/nodes/mention/src/MentionElement/MentionElement.styles.ts
@@ -12,7 +12,7 @@ export const getMentionElementStyles = <V extends Value>(
     {
       root: [
         tw`my-0 mx-px align-baseline inline-block`,
-        props.selected && props.focused && tw`boxShadow[0 0 0 2px #B4D5FF]`,
+        props.selected && props.focused && tw`shadow-[0 0 0 2px #B4D5FF]`,
         css`
           padding: 3px 3px 2px;
           border-radius: 4px;