+++ packages/ui/suggestion/src/PlateSuggestionToolbarButton.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { useSuggestingButton } from '@udecode/plate-suggestion';
+import { ToolbarButton, ToolbarButtonProps } from '@udecode/plate-ui-toolbar';
+
+export const PlateSuggestionToolbarButton = (props: ToolbarButtonProps) => {
+  const buttonProps = useSuggestingButton(props as any);
+
+  return <ToolbarButton {...(buttonProps as any)} />;
+};