+++ packages/ui/nodes/block-quote/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-block-quote-ui
 
+## 20.4.0
+
+## 20.3.2
+
 ## 20.0.0
 
 ## 19.7.0