+++ packages/ui/nodes/code-block/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @udecode/plate-code-block-ui
 
+## 20.4.0
+
+## 20.3.2
+
+## 20.3.1
+
 ## 20.0.0
 
 ### Patch Changes